+++ Tutor_UI/Users/PrijaveTutorForm.cs
@@ -24,7 +24,7 @@ namespace Tutor_UI.Users
 
         private void PrijaveTutorForm_Load(object sender, EventArgs e)
         {
-            
+
             BindNeProcitano();
             BindProcitano();
         }
@@ -47,7 +47,7 @@ namespace Tutor_UI.Users
         private void BindNeProcitano()
         {
             HttpResponseMessage response = banStudentService.GetResponse();
-            
+
 
             if (response.IsSuccessStatusCode)
             {
@@ -57,19 +57,21 @@ namespace Tutor_UI.Users
                 NovePrijaveGridView.DataSource = lstNeProcitanih;
                 NovePrijaveGridView.ClearSelection();
             }
-            
+
         }
 
         private void button1_Click(object sender, EventArgs e)
         {
             //Promjeni nazive tabova
-            
-            int TutorId = 0;                 
-            if (tabControl1.SelectedTab == tabControl1.TabPages["tabPage1"]) {
-                TutorId= Convert.ToInt32(NovePrijaveGridView.SelectedRows[0].Cells[0].Value);
+
+            int TutorId = 0;
+            if (tabControl1.SelectedTab == tabControl1.TabPages["tabPage1"] && NovePrijaveGridView.SelectedRows.Count != 0)
+            {
+                TutorId = Convert.ToInt32(NovePrijaveGridView.SelectedRows[0].Cells[0].Value);
             }
 
-            if (tabControl1.SelectedTab == tabControl1.TabPages["tabPage2"]) {
+            if (tabControl1.SelectedTab == tabControl1.TabPages["tabPage2"] && ProcitanePrijaveGridView.SelectedRows.Count != 0)
+            {
                 TutorId = Convert.ToInt32(ProcitanePrijaveGridView.SelectedRows[0].Cells[0].Value);
             }
 
@@ -93,13 +95,18 @@ namespace Tutor_UI.Users
 
         private void IzbrisiBtn_Click(object sender, EventArgs e)
         {
-            int procitanaPrijava = Convert.ToInt32(ProcitanePrijaveGridView.SelectedRows[0].Cells[0].Value);
-            var response = banStudentService.DeleteResponse(procitanaPrijava.ToString());
+            if (ProcitanePrijaveGridView.SelectedRows.Count != 0)
+            {
+                int procitanaPrijava = Convert.ToInt32(ProcitanePrijaveGridView.SelectedRows[0].Cells[0].Value);
+                var response = banStudentService.DeleteResponse(procitanaPrijava.ToString());
 
-            if (response.IsSuccessStatusCode) {
+                if (response.IsSuccessStatusCode)
+                {
 
-                MessageBox.Show("Uspjenos obrisana prijava");
-                BindProcitano();
+                    MessageBox.Show("Uspjenos obrisana prijava");
+                    BindProcitano();
+                }
+
             }
         }
     }
